% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLNN_GetWts.R
\name{BLNN_GetWts}
\alias{BLNN_GetWts}
\title{Extract the vector of network weights from a BLNN object}
\usage{
BLNN_GetWts(Network)
}
\arguments{
\item{Network}{A BLNN network object}
}
\value{
A vector containing all weight parameters from the network. The weights are sorted
by imput, having all weights from covariate one that connect to hidden node 1:n. Each
subsequent covariates weights for layer one are saved in this order followed by
the weights associated with the bias term for layer one. Layer two weights follow
a similar format having each hidden units weighes followed by bias weights.
}
\description{
Allows the user to extract all weight values from a BLNN network.
object.
}
