% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLNN_Predict.R
\name{BLNN_Predict}
\alias{BLNN_Predict}
\title{Calculate predicted values and errors for a BLNN object}
\usage{
BLNN_Predict(Network, x, y = NULL, fit = NULL)
}
\arguments{
\item{Network}{A BLNN network object}

\item{x}{A set of input data}

\item{y}{A set of target response values. Default is \code{NULL}}
}
\value{
A vector or matrix of predicted responses. If y is given, returns a list containing
\describe{
\item{Errors}{Calculated network error}
\item{Difference}{A vector of differences y-predicted}
\item{Fitted_Values}{A vector or matrix of predicted responses}
}
}
\description{
Allows the user to calculate predicted values from input data with the
option of providing network and fitted errors if known y values are supplied.
}
