% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyBLNN.R
\name{launch_shinyBLNN}
\alias{launch_shinyBLNN}
\title{Launch ShinyStan diagnosis}
\usage{
launch_shinyBLNN(fit)
}
\arguments{
\item{fit}{the BLNN_Train object.}
}
\description{
lanuch the shinyStan tool fo diagnosing the posterior samples using the NUTS algorithm.
}
