% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLNN_UpdateNetwork.R
\name{BLNN_Update}
\alias{BLNN_Update}
\title{Calculate predicted values and errors for a BLNN object}
\usage{
BLNN_Update(NET, fit, index = NULL)
}
\arguments{
\item{fit}{The fitted object returned by \code{\link{BLNN_train}}.}

\item{index}{The starting index used to compute the average estimated value of the newtork weights. The default is warmup+1}

\item{Network}{A BLNN network object}
}
\value{
The Network object with trained weights.
}
\description{
Allows the user to update the network object after examining the samples.
If the samples drawn passed the diagnostic checks then we need to update the network object with the average value of the accetped samples after warmup.
}
